Cherry pick conflict #10995
